chore: refactor formatter and value functions
Refactor the Value functions and the formatter to fix an error with the
formatter when the key was not a string.

The Value* functions were refactored to be more consistent. Now Value*
returnas always the first occurrence and Values* returns all occurrences
of a key.

Now the formatter will print all the keys in the reverse order of
creation (most recent first).

Other formatters were added for convenience.