feat: removes formatter from Error function
The default formatter was causing unwanted behavior with external
wrapping methods, like fmt.Errorf, because of how the Error functions
was implemented and the use of GetRootError.

Some context was lost in those cases.

The solution was to remove de formatter from the Error function. Now the
Error function will only print the inner Error(). This will preserve the
context of other errors.

The formatters are still available to be used directly or by calling
the new Format() function that will apply the formatter set in the
error chain or the Default one. This was the previous behavior of the
Error functions.